requirements.txt
soundfile
numpy
pytaglib
pyinstaller
mutagen